+++ packages/ragbits-chat/CHANGELOG.md
@@ -12,6 +12,9 @@
 - Allow UI cutomization using config endpoint (#643)
 - Add support for live updates and followup messages (#654)
 - Fix invalid context structure in requests from FE (#663)
+- Arrow Up and Arrow Down now cycle through sent messages in a terminal-like style (#667)
+- Fix followup messages not sending (#680)
+- Improve typing of TypeScript libraries (#681)
 
 ## 1.0.0 (2025-06-04)